+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Renamed DAR -> DER form. Renamed Submit Data to ESDIS form to DAR
 - Updated the language for Open Data Policy questions in DAR and DER
 - Removing terraform references to edpub_metrics_s3_bucket
+- Removing unused database tables and references to them
 
 ## 1.1.0